move democracy to old and councils to own file until decided how to migrate councils
